Add generic way to push changes down hierarchy
Add a config property to containers called "propagated_properties".
All PUT updates to these properties will update all objects down
the hierarchy. Update requests that mix propagated_properties with
other properties will not currently be allowed, but modifying the logic
to only push the relevant properties of the payload down the hierarchy
tree is an option. Still to be done: proper permission and schema
validation.